backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pymupdf==1.23.8
python-docx==1.1.0
Pillow==10.2.0
Jinja2==3.1.3
aiofiles==23.2.1
pydantic==2.5.3
reportlab==4.0.8

